+++ js/examples/next-openai-telemetry-app/instrumentation.ts
@@ -1,7 +1,7 @@
 import { registerOTel } from "@vercel/otel";
 import { diag, DiagConsoleLogger, DiagLogLevel } from "@opentelemetry/api";
 import { SimpleSpanProcessor } from "@opentelemetry/sdk-trace-base";
-import { OpenInferenceProtoTraceExporter } from "../../packages/openinference-span-exporter/src";
+import { OpenInferenceProtoTraceExporter } from "../../packages/openinference-trace-exporter/src/OpenInferenceProtoTraceExporter";
 // For troubleshooting, set the log level to DiagLogLevel.DEBUG
 diag.setLogger(new DiagConsoleLogger(), DiagLogLevel.DEBUG);
