+++ js/packages/openinference-trace-exporter/src/OpenInferenceHttpTraceExporter.ts
@@ -7,19 +7,6 @@ import { OTLPExporterNodeConfigBase } from "@opentelemetry/otlp-exporter-base";
 import { ReadableSpan } from "@opentelemetry/sdk-trace-base";
 import { ExportResult } from "@opentelemetry/core";
 
-export class OpenInferenceProtoTraceExporter extends ProtoExporter {
-  constructor(config?: OTLPExporterNodeConfigBase) {
-    super(config);
-  }
-  async export(
-    spans: ReadableSpan[],
-    resultCallback: (result: ExportResult) => void,
-  ) {
-    const openInferenceSpans = addOpenInferenceAttributesToSpans(spans);
-    return super.export(openInferenceSpans, resultCallback);
-  }
-}
-
 export class OpenInferenceHttpTraceExporter extends HttpExporter {
   constructor(config?: OTLPExporterNodeConfigBase) {
     super(config);
@@ -33,17 +20,3 @@ export class OpenInferenceHttpTraceExporter extends HttpExporter {
     return super.export(openInferenceSpans, resultCallback);
   }
 }
-
-// export class OpenInferenceGrpcTraceExporter extends GrpcExporter {
-//   constructor(config?: OTLPExporterNodeConfigBase) {
-//     super(config);
-//   }
-
-//   async export(
-//     spans: ReadableSpan[],
-//     resultCallback: (result: ExportResult) => void,
-//   ) {
-//     const openInferenceSpans = addOpenInferenceAttributesToSpans(spans);
-//     return super.export(openInferenceSpans, resultCallback);
-//   }
-// }
