+++ js/packages/openinference-trace-exporter/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./OpenInferenceGrpcTraceExporter";
+export * from "./OpenInferenceHttpTraceExporter";
+export * from "./OpenInferenceProtoTraceExporter";
