+++ js/packages/openinference-trace-exporter/src/utils/typeUtils.ts
